+++ src/components/Editor/EditorContent/EditorContent.tsx
@@ -11,7 +11,7 @@ import useEditorStore from "@/store/useEditorStore";
  * 드래그&드롭, 실행취소 기능을 포함한 에디터 인스턴스를 생성
  */
 const EditorContent = memo(() => {
-  const { setEditor } = useEditorStore();
+  const setEditor = useEditorStore((state) => state.setEditor);
   const editorInstanceRef = useRef<EditorJS | null>(null);
 
   useEffect(() => {
